+++ crates/test/src/lib.rs
@@ -128,3 +128,5 @@ pub use __rt::criterion::Criterion;
 
 // web_time Instant
 pub use __rt::web_time::Instant;
+
+pub use __rt::browser::screenshot;